server/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pillow>=10.0.0
pytesseract>=0.3.10
